[ADD] Add size and color parameter to TextDrawing
In the future versions of the firmware, it will be possible to choose the color and the size of a text. To help developers to prepare this feature, the API has been completed with these two parameters. However, do not be surprised that it does not impact the color or the size of the element on the aRdent glasses.