Convert WhiteScreen into BlankScreen with a color parameter  (#18)
This PR adds the possibility to change the background color of the screen.
To do so,  `WhiteScreen` objects are replaced by `BlankScreen` with a color parameter. This parameter (optional) describes the color to use as background. If not specified, it remains the same as before.

Note : This feature will only work from firmware 1.1.5

---------

Co-authored-by: Antoine Malherbe <a.malherbe@getyourway.be>