Add a method to read the firmware version of the device (#22)
Add a new method that reads the firmware version of the device. The value returned by the function is a `FirmwareVersion` object that has 3 integers, the major, the minor, and the patch versions. `FirmwareVersion` can be compared between them using comparison operators.

The firmware version is read from the firmware revision string characteristic from the following service:
https://www.bluetooth.com/specifications/specs/device-information-service/

---------

Co-authored-by: Antoine Malherbe <a.malherbe@getyourway.be>