Support custom icons (#29)
Some use cases require custom icons that are not necessarily part of the library. Before this change, theses custom icons blocked the entire deserialization process. They are now supported, and the icon will be sent correctly to the glasses.
This change also fixes the problem of an app that is not synchronous with the last icon library updates.