Add a new drawing for colored rectangles which allows for partially clearing the screen (#38)
The new drawing is `RectangleDrawing` which allows for drawing colored rectangles. If the color is null, the rectangle will use the current background color which can be used to partially clear the screen.

A list with a short description of every drawing that can be sent has been added to the documentation.