Remove some technical features that are not needed in the library (#78)
Some features are seen as advanced and quite technical. They are usually not needed to use the glasses. In order to keep a simple and easily maintainable library, these have been removed from the library.

Those methods include:
* `setBrightness`, `setContrast`
* `enableBacklight`, `autoRotateScreen`

The version has also been bumped to 2.0.2 and the package dependencies have been upgraded.